# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "B2C Marketing Automation API",
    "description": "Complete REST API for the B2C Marketing Automation Platform. Manage contacts, workflows, deliveries, segments, audiences, custom objects, and more.",
    "version": "2.0.0",
    "contact": {
      "name": "Platform Team"
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "Current server"
    }
  ],
...
  "paths": {
...
    "/contacts": {
      "get": {
        "tags": ["Contacts"],
        "summary": "List contacts",
        "description": "Get all contacts with pagination and filtering support.",
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "default": 1 } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "default": 50 } },
          { "name": "status", "in": "query", "schema": { "type": "string" } },
          { "name": "search", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated list of contacts" } }
      },
      "post": {
        "tags": ["Contacts"],
        "summary": "Create contact",
        "description": "Create a new B2C consumer profile.",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactInput" } } } },
        "responses": { "201": { "description": "Contact created" } }
      }
    },
...
    "/contacts/{id}": {
      "get": {
        "tags": ["Contacts"],
        "summary": "Get contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Contact details" }, "404": { "description": "Not found" } }
      },
      "put": {
        "tags": ["Contacts"],
        "summary": "Update contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ContactInput" } } } },
        "responses": { "200": { "description": "Contact updated" } }
      },
      "delete": {
        "tags": ["Contacts"],
        "summary": "Delete contact",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Contact deleted" } }
      }
    },
...
    "/segments": {
      "get": {
        "tags": ["Segments"],
        "summary": "List segments",
        "responses": { "200": { "description": "List of segments" } }
      },
      "post": {
        "tags": ["Segments"],
        "summary": "Create segment",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SegmentInput" } } } },
        "responses": { "201": { "description": "Segment created" } }
      }
    },
...
    "/segments/{id}": {
      "get": {
        "tags": ["Segments"],
        "summary": "Get segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment details" } }
      },
      "put": {
        "tags": ["Segments"],
        "summary": "Update segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment updated" } }
      },
      "delete": {
        "tags": ["Segments"],
        "summary": "Delete segment",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": { "200": { "description": "Segment deleted" } }
      }
    },
...
    "/audiences/{id}": {
      "get": { "tags": ["Audiences"], "summary": "Get audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Audience details" } } },
      "put": { "tags": ["Audiences"], "summary": "Update audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Updated" } } },
      "delete": { "tags": ["Audiences"], "summary": "Delete audience", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Deleted" } } }
    },
    "/audiences/{id}/members": {
      "get": { "tags": ["Audiences"], "summary": "Calculate audience members", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Member list" } } }
    },
    "/audiences/{id}/import-csv": {
      "post": { "tags": ["Audiences"], "summary": "Import contacts from CSV", "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }], "responses": { "200": { "description": "Import results" } } }
    },
...
  },
  "components": {
    "schemas": {
      "ContactInput": {
        "type": "object",
        "properties": {
          "email": { "type": "string", "format": "email" },
          "first_name": { "type": "string" },
          "last_name": { "type": "string" },
          "phone": { "type": "string" },
          "status": { "type": "string", "enum": ["active", "inactive"] },
          "city": { "type": "string" },
          "state": { "type": "string" },
          "country": { "type": "string" }
        },
        "required": ["email", "first_name", "last_name"]
      },
...
      "SegmentInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "segment_type": { "type": "string" },
          "rules": { "type": "array", "items": { "type": "object" } }
        },
        "required": ["name"]
      },
...
    }
  }
}